requirements.txt
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.7.0
flake8==6.1.0
mccabe==0.7.0
psycopg2-binary==2.9.10
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
